feet-last

+++ src/main/java/com/badiagroup/Library/repository/loans_repository.java
@@ -1,9 +1,13 @@
 package com.badiagroup.Library.repository;
 
 import com.badiagroup.Library.model.loans;
+
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface loans_repository extends JpaRepository<loans, Integer> {
+public interface loans_repository extends JpaRepository<loans, Long> {
 }
